+++ Qase.Csharp.Commons/config/ExternalLinkType.cs
@@ -1,0 +1,18 @@
+namespace Qase.Csharp.Commons.Config
+{
+    /// <summary>
+    /// Enumeration for external link types
+    /// </summary>
+    public enum ExternalLinkType
+    {
+        /// <summary>
+        /// Jira Cloud integration
+        /// </summary>
+        JiraCloud = 1,
+
+        /// <summary>
+        /// Jira Server integration
+        /// </summary>
+        JiraServer = 2
+    }
+}